Update freebsd.rb
In one of these lines lies the error, which leads to the put in "_enable" twice! As a result, the necessary service certainly does not start. Unfortunately, in ruby I do not understand and can not reliably determine what is wrong: in a regular expression that specifies the name of a variable or a method of editing the variable file.